Adding keyterm extraction from text.
This commit add the functionality to extract keyterms from text based on
both frequency and influence (Pearsons Chi-Squared). In the course of
adding this functionality some changes were made to index and frames.

A global count of all words per frame was added to the index to make it
easier to make the calculations required to extract keyterms. Also, a
set of unique terms was added to each frame object at index time for the
same reason.

Some package renaming was done to cleanup the source tree. The package
analysis was renamed to processing and a new package called analytics
was created.

Finally, numpy was removed as a requirement from setup.py and it is no
longer being used (although it might be used again in the future).

Issue: http://agilezen.com/project/54516/story/3
Change-Id: Ib37f1f83748adee45984cbc0bfa24a18fb90f999