Added a new analyser design.
The tokenization and filtering of text is now handled by an analyser
framework. If you want to tokenise text you need to call a pre-existing
analyser or define your own. Analysers use tokenizers to get individual
tokens before passing them through 0 or more filters. Each filter has
the chance to look at the token and take some action. What you're left
with is a tokenized and filtered version of your text.

The new famewrok is based completely on generators and is very careful
about the amount of objects it creates for performance. Both index and
bi-gram discovery have been moved to the analyser framework. The use of
the framework for bi-gram is quite dirty mainly to make it as fast as
possible.

Also added:
* A .coveragerc to make the generation of html converage reports a
little more standard.
* A bi_grams example.

Issue: http://agilezen.com/project/54516/story/41
Change-Id: I126fd51a45562283ef24033e13c0ac91a6ef0285