Fixing bug introduced in Story41.
Bi-gram filter had a bug which meant the first token of a bi-gram was
yielded when it shouldn't of been leading to a double count. For
example, 'golden' appears in alice 7 times where 'golden key' appears 6
times. With this bug, the count of 'golden' remained at 7 when it should
actually be 1.

Also added tests for this specific issue to make sure it doesn't happen
again.

Change-Id: Ia466623d3a9d1dd7f6150d36914bd699a89f8d37