Storage framework with a RamStorage implementation
Implementing this story has meant changes to almost all parts of the
code. It has spawned a new plugin system through which the influence
analytics are now implemented. The topic identification from the layout
code has also been moved into that same analytics plugin.

A consequence of implementing this story was to fix a bug with word
tokenisation where a name at the end of a sentence would get compunded
together with the first word of the next sentence. Instead, this will
now only happen if the word is between 1 & 3 charactes long (for example
K. Rogers, Mr. Rogers and Mrs. Rogers).

Another consequence was the discover of a bug in the old influence code
where it was using an incorrect associations table when case folding had
been performed on the index.

The new influence code departs from the previous code in that it now
stores all influence relationships (including negative ones). An
influence threshold is now only considered when calculating cumulative
influence. It is hoped the recording of negative influence might help
with sentiment and other endevours.

This is partial implementation of Issue 40 but enough to close the
issue. An SQLite Storage implementation will be next but will create a
separate issue for it.

Finally, my terminal has no spell checker so this commit message may be
bugged.

Issue: http://agilezen.com/project/54516/story/40
Change-Id: I888f47a851787b73bac42e3b205eaab4ca85629c