Adding support for index update.
Updating an index means only adding frames to the index which have yet
to be added. It differes from a reindex in that it doesn't clear the
index for a complete rebuild.

As part of this change we have moved to storing the analysed metadata
values directly on the frames at _metadata. This is to avoid having to
re-do the analyse computation for categorical fields.

Also, an _indexed field has been added to frames. It should be pretty
self explanitory....

Issue: http://agilezen.com/project/54516/story/65
Change-Id: I4f551b5f75668ef7ff5cd731a360be1c2493e60b