Adding an info container to the Index.
As part of this change versioning of the index and and attempt at state
saving has been attempted. Versioning is storing a variable in the
container which is meant to be synonomous to a changeset id. You should
be able to store a resvision number and compare it to the one on the
index to tell if the version you have is up-to-date. This should be
useful for caching.

In an attempt to detect if an index is up-to-date a clean clean flag has
also been added tot he info container refelcting if the index data
structures are up to date or not.

A side affect of this changeset has been to streamline test-running.
Undesecary uses of slow storage backend have been removed from places
where they aren't required, resulting in a 100 second reduction in run
time on my machine.

Issue: http://agilezen.com/project/54516/story/55
Change-Id: Ibe8b036c0b8f2b7add117f70b5e09e8f30d04a8f