Bug fixes as a result of Chrysalis story21
While implementing Chrysalis story21 and trying to use caterpillar as a datastore for user credentials I identified and fixed a number of caterpillar issues. They include:

* The handling of documents with only categorical fields whether they have indexed fields or not.
* Handling of metadata when reindexing in update mode.
* Lack of tests for various corner cases.
* Numerous pep8 fixes.

With these fixes caterpillar can now be used a stand-alone data store for the datatypes it supports. A date field type still needs to be added.

Change-Id: I29d14d8c0aeb18e781bc8721f06b711256b51c96