Defaulting to excel csv dialect.
The auto dialect detection in python csv seems to be useless. It has
been observed to fail to recognise the presence of double quotes in a
csv file thus completely distorting the column structure. The same file
opens fine in OpenOffice and parses fine when we default to excel
dialect rather than automatic dialect.

Change-Id: I94436e32500b53aaf2401316d03871f4bdc78369