Fixed bugs with topic identificatoin.
Firstly, merge topics is sensitive to sort ordering. So we have to
maintain deterministic ordering of topics in this method. This is due to
the fact that sometimes multiple merge operations have equal candidacy,
    the one being picked only a result of topic order.

Also, found a regression in merging due to changing the flow control
when refactoring to the topics plugin. This was causing modification to
topics without updating the list of topics being iterated over. This is
a likely candidate for a race condition. I was able to consistently
reproduce topic bugs by running the whole coverage tests at once, which
will obviously slow down the execution of the tests and provide an
opportuntiy for a race condition to produce different results.

Also updated the tests to more thoroughly test topics.

Change-Id: I55304bda562b6329284d6fe0eb57af46c87cbeed