New whitespace word tokenizer and punctuation filter
Tokenisation of words now defaults to a tokeniser that splits on
whitespace and captures compound names. The default analyser now also
uses a filter that allows certain leading and trailing punctuation and a
filter that removes possessive contractions in a really robust manner.

There were a bunch of changes to test that were really annoying. I've
now created test analyers that should change ever. When new tokenisers
and analysers are added, we should only ahve to add tests to the
test_analyse etc., not adjust all tests. Of the tests that were updated:

* test_index_merge_terms: An instance of party became tea-party
* test_index_moby_case_folding: Multiple cases of terms being joined to
* other terms because of punctuation. For example - 'Flask--good-bye'
* test_searching_reserved_words: Multiple cases of 'and' joined to other
* terms with punctuation
* test_index_alice_bigram_discovery: The two missing bigrams ('beau
  ootiful' and 'guinea pigs') and now single terms joined with 1
* or more '-'.
* test_searching_alice: Two places where a *ing ending term was joined
* to another term with 1 or more '-'.
* test_searching_alice: One case of 'voice' term joined with '-'.
* test_index_alice: All differences in vocab size verified by hand.

Also fixed a small bug with setup.py.