Per field frame representation for IndexWriter
The representation of frames for indexed text changes from:
	{frame_id: {frame_content}} to
	{field: {frame_id: {frame_content}}

This allows per field indexing (and eventually searching) of text.
This change flows on to all of the IndexWriter in memory index
representation (frequencies, associations, positions).

There is now a set of 4 tables (frequencies, assocs, positions,
frames) for each indexed text field, with the table name
type_field_name in the storage layer. Separate tables are
necessary for effective searching by frame_id.