Redefine the plugin interface
A plugin is now an instance of an AnalyticsPlugin class and is
identified by a unique combination of settings and plugin name.

Plugins need to implement a set of methods that allow serialisation
of settings and state.

Plugin data is centralised in two tables: a plugin registry with
one row per combination of plugin name and plugin settings, and a
plugin data registry which is a key value store for plugin data.

Plugins are responsible for serialising and deserialising their own
state, as well as consistently serialising their settings so the
plugin registry can be reliably examined.