Clear up plugin API by renaming get_name() to get_type()
Plugins are now identified by a combination of a string
``plugin_type`` and the corresponding settings for a plugin
instance. The plugin_type string is returned by the method
get_type() of the AnalyticsPlugin class, replacing the get_name()
method.

This avoids the confusion from calling it get_name(), where it
isn't clear that the plugin_name is actually a name for a family
of plugins, not a unique plugin instance.