Allow APSW to retry on busy before raising error
Mitigates #78 by setting the busytimeout for the SQLite connection
object. This means that if database is locked to cleanup the WAL
file the connection will retry for up to 1s before raising an error.

This fix should lessen the impact of cleaning up the WAL file
requiring an exclusive lock, however it can't guarantee that the
problem won't occur, in particular errors will still occur if
cleaning up the WAL file requires a long time.