Make indexing minimally invasive
Replace destructive fold_term_case and merge_terms on a writer with the following:
	- reader.get_case_fold_terms(): for suggesting case fold merges
	- reader.detect_significant_ngrams(): for finding signficiant ngrams on the index
	- reader.search_ngrams(): for finding frames containing ngrams on an index

Currently only bigrams are supported for the ngram functions.

In order to efficiently support the ngram cases term positions are now encoded as
bitstrings of the term's ordinal position in a frame. This allows fast set comparison
options, and is exact for frames up to 63 words in size. Currently frames larger than
this are encoded with modulo positions and marked as approximate, but this is only
currently used to ignore that frame.

DefaultAnalyser has been changed: the stopword list is now empty, and all size words
are included by default.

Adds a new iterator for term-frequency vectors of frames. This allows easier connection
to scipy for calculation of frame-term and association matrices. This is also a starting
point for allowing modelling operators to operate on subsets of frames (such as from a
query).

The destructive operations fold_term_case and merge_terms have been completely removed,
consequently there are updates to all test examples that relied on the counts for these.

Notes

Currently only bigrams are supported for ngram searching.