Remove object oriented search API
Remove all components of the search API, including the 10+ classes
used to implement search functionality. This is a significant
maintenance burden: keeping the search functionality in sync with
the IndexReader functionality is non trivial.

Furthermore, arbitrary scoring is non trivial from a performance
and indexing point of view: without support for additional data
stored in the index most alternative scorers are not implementable.
This change takes the opionated view that we should move scoring
to implementations in the storage layer.

To replace it, the IndexReader method now provides efficient low
level search and filter operations that execute in the data layer,
and utility functions for composition of resultsets.

Result sets are encapsulated as dictionaries of frame/document_ids
and associated sets of scores, allowing for simple and efficient
set operations and simple modification as necessary. This promotes
interoperability with other systems: a set of frame_ids from say
a topic model can be directly integrated with filter results from
caterpillar with minimum effort.

This does remove some functionality that we might need to look at
adding back in later:
	- Term weighting (easy)
	- Wildcard term matching (difficult - needs low level support to
	  avoid walking the entire vocabulary for matches)
	- High level search through a string API - QueryStringQuery can
	  eventually be repurposed for this, with the leaf nodes of the
	  query mapping to calls to IndexReader.filter().